Fix wxSlider ticks appearance in wxOSX
To have a proper centering with respect to the slider's line, the size
of the slider in the dimension perpendicular to the slider's line has to
be an even number. Therefore,when determining the best size, the size is
adjusted to an even number when ticks are shown.

See #24532, #24537.

(cherry picked from commit a4a334e7f7029be45283302a8342ff7144e1cf36)