Use ISO 8601 in wxDatePickerCtrl if current locale can't be used
Some locales use format specifiers not supported (yet) by
wxDateTime::Parse() and while the best fix is to add support for them,
as it has been already done for %F (see #24173) and will be done for
"%-" (see #24533) soon, it still seems wise to handle the locales for
which it still might happen, now or in the future, more gracefully by
falling back to ISO 8601 date format if the locale format can't be used
because we can't parse it.

Note that an alternative approach could consist in using the locale
format for formatting the dates and falling back to ISO 8601 only for
parsing them, but it would arguably be confusing to the user if the
program expected input in a format different from what it uses for
displaying the dates, so prefer just using the same format for both, if
we can't do otherwise.

(cherry picked from commit 8e66698d273b51d0b32333a69473511c7c14bba4)