Support padding format flags in wxDateTime::ParseFormat()
This is not really used when parsing, but we still need to support the
GNU extensions to the strftime() formatting specifiers used to change
padding, as they can (and do, at least in the case of "%-d" and "%-m")
occur in the locale descriptions and so can be used when parsing dates
using the current locale format.

See #24533.

(cherry picked from commit 2cc3a3814ce43deaf7a43926870c5725d449415e)